Dev (#47)
* T&C (#16)

* chore: added tos

* chore: fixed spacing

and link to tos

* fix: added page title

* chore: tos spacing

* chore: privacy policy

* chore: added padding to tos & pp pages

* fix: pp date

* Filters and chores (#18)

* fix: map route

* feat: basic ui for filters

* feat: combobox improvements

* feat: multiple selector filter

* feat: connect inputs to state

* fix: minor bugs

* feat: filters filter

* fix: search and filter bugs

* chore: added loading state to tag search

* chore: bounty page ssg

* chore: org page ssg

* chore: increased loading message

* chore: improved loading state for submissions

in bounty page
+ hide submissions tab if user is not logged in

* chore: improved loading state for bounty page

show an animation when loading the org name and if the user can submit


* Misc issues (#20)

* fix: mapFilter, newMsg, navbar, slowness(may be)

* fix: adding back country

* feat: ssg map data

+ changed .env. Reading directory instead of doing a http request
+ updated prisma package
+ map data ssg 

* chore: ssg orgs page

* test: trying tRPC prefetch (did not work)

+ installed sharp as next suggested

* chore: style tweaks


* fix: adding sanction and discord link (#22)

* fix: build fix

* Adding link (#25)

* fix: adding sanction and discord link

* fix: removing check on optionals and not saving empty values

* Adding link (#27)

* fix: adding sanction and discord link

* fix: removing check on optionals and not saving empty values

* fix: cover image

* Mobile UI (#30)

* fix: mobile UI

* fix: invoice payment error logs

* fix: submissions list limit

* Mobile UI (#32)

* fix: mobile UI

* fix: invoice payment error logs

* fix: submissions list limit

* feat: wallet connection in mobile

* fix: url correction

* fix: minor fix

* fix: mobile UIs after logIn

* fix: notif icon gap

* Subs pagination and rewards (#34)

* fix: pagination of submissions in bounty and accounts

* fix: rewards page and filters in subs and invoices

* fix: pagination in invoices and rewards

* fix: map color correction

* fix: show rewards only of the user

* Subs pagination and rewards (#36)

* fix: pagination of submissions in bounty and accounts

* fix: rewards page and filters in subs and invoices

* fix: pagination in invoices and rewards

* fix: map color correction

* fix: show rewards only of the user

* fix: reviews page

* fix: correct wordings

* New bounties (#39)

* chore: add new bounties

* fix: removing components that are not implemented and correcting title

* fix: adding TODO

* fix: correcting type

* fix: handling newlines and double quotes

* fix: notif content fix

* New bounties (#40)

* chore: add new bounties

* fix: removing components that are not implemented and correcting title

* fix: adding TODO

* fix: correcting type

* fix: handling newlines and double quotes

* fix: notif content fix

* fix: log path

* fix: admin log folder

* fix: removing log

* fix: logging in console

* feat: bounty edit and delete features

* fix: adding admin menu item

* chore: targets page

* fix: tags in bounty list and target bio in bounty

* Bounty for names (#42)

* fix: name bounty type and none

* fix: text and styling fixes

* test: added pre-build cmd

* test: added vercel-build cmd


* Adding revalidate (#44)

* fix: name bounty type and none

* fix: text and styling fixes

* test: added pre-build cmd

* test: added vercel-build cmd

* fix: revalidate


* Bounty increment (#46)

* fix: name bounty type and none

* fix: text and styling fixes

* test: added pre-build cmd

* test: added vercel-build cmd

* fix: revalidate

* fix: structure correction

* chore: schedule cron job

* fix: file location

* feat: update new bounty config

* feat: edit wallet control of bounty

---------


---------

+++ src/server/routes/bounties/deleteBounty.ts
@@ -1,3 +1,5 @@
+import { BountyState } from '@prisma/client';
+import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 
 import { LOG } from '@/server/importer/utils';
@@ -12,10 +14,27 @@ export type DeleteBountyParams = z.infer<typeof DeleteBountySchema>;
 export const deleteBounty = staffOnlyProcedure
   .input(DeleteBountySchema)
   .mutation(async ({ input, ctx: { prisma } }) => {
-    const bounty = await prisma.bounty.delete({
+    const bounty = await prisma.bounty.findUnique({
       where: {
         slug: input.slug,
       },
     });
-    LOG(`${bounty.slug}: Deleted bounty "${bounty.title}"`);
+    if (!bounty) {
+      throw new TRPCError({
+        code: 'BAD_REQUEST',
+        message: 'Bounty not found',
+      });
+    }
+    if (bounty?.status === BountyState.Closed) {
+      throw new TRPCError({
+        code: 'FORBIDDEN',
+        message: 'Cannot delete closed bounty',
+      });
+    }
+    const deletedBounty = await prisma.bounty.delete({
+      where: {
+        slug: input.slug,
+      },
+    });
+    LOG(`${bounty?.slug}: Deleted bounty "${bounty?.title}"`);
   });
